+++ libraries/dagster-hex/dagster_hex_tests/test_hex.py
@@ -1,5 +1,3 @@
-import dagster as dg
-
 from dagster_hex.resources import HexResource
 
 
@@ -26,7 +24,7 @@ def test_run_project(requests_mock):
     assert response == {"data": "mocked response"}
     assert requests_mock.last_request.json() == {
         "inputParams": {"param": "var"},
-        "updateCache": False,
+        "useCachedSqlResults": True,
     }
 
 
@@ -41,32 +39,5 @@ def test_run_project_no_input(requests_mock):
     response = hex.run_project("abc-123")
     assert response == {"data": "mocked response"}
     assert requests_mock.last_request.json() == {
-        "updateCache": False,
-    }
-
-
-def test_run_project_from_asset(requests_mock):
-    project_id = "hex-project-id"
-    requests_mock.post(
-        f"https://testurl/api/v1/project/{project_id}/run",
-        headers={"Content-Type": "application/json"},
-        json={"data": "mocked response"},
-    )
-
-    @dg.asset()
-    def example_hex_project_asset(hex: HexResource) -> dg.MaterializeResult:
-        response = hex.run_project(project_id, inputs={"param": "var"})
-        return dg.MaterializeResult(metadata={"data": response.get("data")})
-
-    res = dg.materialize(
-        [example_hex_project_asset],
-        resources={
-            "hex": HexResource(api_key="HEX_API_KEY", base_url="https://testurl/")
-        },
-    )
-    assert res.success
-
-    assert requests_mock.last_request.json() == {
-        "inputParams": {"param": "var"},
-        "updateCache": False,
+        "useCachedSqlResults": True,
     }