+++ libraries/dagster-hex/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 ---
 
+## 0.1.5 - UNRELEASED
+### Added
+
+- Hex project asset definition with `build_hex_asset` function using the project_id
+- Update to hex resource to sync with new hex api updates to deprecate `update_cache` and add `use_cached_sql` instead.
+
 ## 0.1.4 - UNRELEASED
 
 ### Added